had to convert mp3 files to wav files
This was a problem before already, but I forgot to previously do it
in this script. The files from the itps corpus are called wav files
but are actually mp3 files, so we have to convert them. I am converting
them with ffmpeg through the command line here, because it was easiest
this way